minor: move to reflection
